fixed surfacebuilders and removed several unneeded mixins that supported the old way
things to do as a result of this change:

Make large dimstone patches be an ore or ore-like feature so you can control frequency and size easier instead of relying on surfacebuilder rng voodoo.

Make glowdust layering be a feature similar to cinderscape's ashes or vanilla's snowandice feature. This way, the glowdust gets placed correctly on slopes and wont be floating above carvers/caves/pits that carved out the below land.